+++ clients/admin-ui/cypress/e2e/config-wizard.cy.ts
@@ -7,10 +7,6 @@ describe("Config Wizard", () => {
     cy.intercept("PUT", "/api/v1/organization**", {
       fixture: "organization.json",
     }).as("updateOrganization");
-
-    cy.intercept("POST", "/api/v1/generate", {
-      fixture: "generate/system.json",
-    }).as("postGenerate");
   });
 
   beforeEach(() => {
@@ -44,19 +40,47 @@ describe("Config Wizard", () => {
       // Select AWS to move to form step.
       cy.getByTestId("add-system-form");
       cy.getByTestId("aws-btn").click();
+      // Fill form
+      cy.getByTestId("authenticate-aws-form");
+      cy.getByTestId("input-aws_access_key_id").type("fakeAccessKey");
+      cy.getByTestId("input-aws_secret_access_key").type("fakeSecretAccessKey");
+      cy.getByTestId("input-region_name").type("us-east-1{Enter}");
     });
 
     it("Allows submitting the form and viewing the results", () => {
-      cy.getByTestId("authenticate-aws-form");
-      cy.getByTestId("input-aws_access_key_id").type("fake-access-key");
-      cy.getByTestId("input-aws_secret_access_key").type(
-        "fake-secret-access-key"
-      );
-      cy.getByTestId("input-region_name").type("us-east-1{Enter}");
-      cy.getByTestId("submit-btn").click();
+      cy.intercept("POST", "/api/v1/generate", {
+        fixture: "generate/system.json",
+      }).as("postGenerate");
 
+      cy.getByTestId("submit-btn").click();
       cy.wait("@postGenerate");
+
       cy.getByTestId("scan-results-form");
+    });
+
+    it("Displays API errors and allows resubmission", () => {
+      cy.intercept("POST", "/api/v1/generate", {
+        statusCode: 403,
+        body: {
+          detail: "The security token included in the request is invalid.",
+        },
+      }).as("postGenerate403");
+      cy.getByTestId("submit-btn").click();
+      cy.wait("@postGenerate403");
+      // Expect the custom message for this specific error.
+      cy.getByTestId("scanner-error");
+      cy.getByTestId("permission-msg");
+
+      cy.intercept("POST", "/api/v1/generate", {
+        statusCode: 500,
+        body: "Internal Server Error",
+      }).as("postGenerate500");
+      cy.getByTestId("submit-btn").click();
+      cy.wait("@postGenerate500");
+      // Expect the generic message with a log.
+      cy.getByTestId("scanner-error");
+      cy.getByTestId("generic-msg");
+      cy.getByTestId("error-log").contains("Internal Server Error");
     });
   });
 });