+++ clients/admin-ui/src/features/data-qualifier/data-qualifier.slice.ts
@@ -2,8 +2,7 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 
 import type { AppState } from "~/app/store";
-
-import { DataQualifier } from "./types";
+import { DataQualifier } from "~/types/api";
 
 export interface State {
   dataQualifier: DataQualifier[];
@@ -20,28 +19,28 @@ export const dataQualifierApi = createApi({
   }),
   tagTypes: ["Data Qualifier"],
   endpoints: (build) => ({
-    getDataQualifier: build.query<DataQualifier, void>({
+    getAllDataQualifiers: build.query<DataQualifier[], void>({
       query: () => ({ url: `data_qualifier/` }),
       providesTags: () => ["Data Qualifier"],
     }),
   }),
 });
 
-export const { useGetDataQualifierQuery } = dataQualifierApi;
+export const { useGetAllDataQualifiersQuery } = dataQualifierApi;
 
 export const dataQualifierSlice = createSlice({
   name: "dataQualifier",
   initialState,
   reducers: {
-    setDataQualifier: (state, action: PayloadAction<any>) => ({
+    setDataQualifiers: (state, action: PayloadAction<any>) => ({
       ...state,
       dataQualifier: action.payload,
     }),
   },
 });
 
-export const { setDataQualifier } = dataQualifierSlice.actions;
-export const selectDataQualifier = (state: AppState) =>
+export const { setDataQualifiers } = dataQualifierSlice.actions;
+export const selectDataQualifiers = (state: AppState) =>
   state.dataQualifier.dataQualifier;
 
 export const { reducer } = dataQualifierSlice;