+++ clients/admin-ui/src/features/taxonomy/DataCategoryChecklist.tsx
@@ -1,9 +1,10 @@
 import { Box, Heading, SimpleGrid, Text } from "@fidesui/react";
 import { useMemo, useState } from "react";
 
+import { DataCategory } from "~/types/api";
+
 import CheckboxTree from "../common/CheckboxTree";
 import { transformDataCategoriesToNodes } from "./helpers";
-import { DataCategory } from "./types";
 
 interface Props {
   dataCategories: DataCategory[];