+++ clients/admin-ui/src/features/taxonomy/TaxonomyTabContent.tsx
@@ -1,0 +1,31 @@
+import { Spinner, Text } from "@fidesui/react";
+
+interface TaxonomyEntity {
+  fides_key: string;
+  name?: string;
+  description?: string;
+}
+
+interface Props {
+  isLoading: boolean;
+  data: TaxonomyEntity[] | undefined;
+}
+const TaxonomyTabContent = ({ isLoading, data }: Props) => {
+  if (isLoading) {
+    return <Spinner />;
+  }
+  if (!data) {
+    return <Text>Could not find data.</Text>;
+  }
+
+  // TODO: Build actual component, just render data simply for now (#853)
+  return (
+    <>
+      {data.map((d) => (
+        <Text key={d.fides_key}>{d.name}</Text>
+      ))}
+    </>
+  );
+};
+
+export default TaxonomyTabContent;