+++ clients/admin-ui/src/features/taxonomy/types.ts
@@ -1,9 +1,3 @@
-import { FidesBase, FidesKey } from "../common/fides-types";
-
-export interface DataCategory extends FidesBase {
-  parent_key: FidesKey | null;
-}
-
 export interface DataCategoryNode {
   value: string;
   label: string;